wl: Encapsulate CogWlDisplay create/destroy logic
Wrap the creation and destruction of a Wayland display inside of the
cog_wl_display_create() and cog_wl_display_destroy() functions and move
these to cog-utils-wl.
