+++ platform/wayland/meson.build
@@ -133,6 +133,7 @@ wayland_platform_plugin = shared_module('cogplatform-wl',
     'cog-im-context-wl.c',
     'cog-platform-wl.c',
     'cog-popup-menu-wl.c',
+    'cog-utils-wl.c',
     'cog-view-wl.c',
     'os-compatibility.c',
     wayland_platform_sources,
